backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
pydantic==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
llama-index==0.9.11
openai==1.3.5
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
tqdm==4.66.1
pymupdf==1.23.7
